Add Deeplink hook
